Code cleanup
- Removed harmony patch for ModuleAutoCutDrogue from the module's cs file and moved it to its own harmony patches folder
