+++ Source/KSPCommunityPartModules.csproj
@@ -34,6 +34,7 @@
     <PackageReference Include="KSPBuildTools" Version="0.0.2-alpha.7" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="HarmonyPatches\ModuleParachutePatch.cs" />
     <Compile Include="Modules\ModuleAutoCutDrogue.cs" />
     <Compile Include="Modules\ModuleCoPFollowTransform.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
